Fix CI failures and CodeRabbit findings from PR #142
- e2e: fix int32/int64 type mismatch in soak_mode_test assertions (caused 30s timeout)
- helm-crds-check: sync soakMode + recentVerdicts to Helm chart CRDs; add MaxItems=100 marker (#1/#5)
- dashboard: guard against undefined request in submitAgentRequestVerdict (#2)
- gateway: replace CONFLICT string-matching with errVerdictWrongPhase sentinel error (#4)
- accuracy controller: handle AlreadyExists separately inside RetryOnConflict (#6)
- accuracy controller: return error on transient failure instead of silently dropping verdict (#7)
- accuracy controller tests: isolate each test with its own agent identity (#8)
- agentrequest controller: return error on non-NotFound GR lookup failure (#9)
- agentrequest controller: Inc agentRequestActive for PhaseAwaitingVerdict to prevent gauge underflow (#10)

+++ test/e2e/soak_mode_test.go
@@ -159,8 +159,8 @@ var _ = Describe("SoakMode and Accuracy Tracking", Ordered, func() {
 			var summary governancev1alpha1.DiagnosticAccuracySummary
 			err := k8sClient.Get(ctx, types.NamespacedName{Name: summaryName, Namespace: ns}, &summary)
 			g.Expect(err).NotTo(HaveOccurred())
-			g.Expect(summary.Status.TotalReviewed).To(Equal(int32(1)))
-			g.Expect(summary.Status.CorrectCount).To(Equal(int32(1)))
+			g.Expect(summary.Status.TotalReviewed).To(Equal(int64(1)))
+			g.Expect(summary.Status.CorrectCount).To(Equal(int64(1)))
 			g.Expect(summary.Status.DiagnosticAccuracy).NotTo(BeNil())
 			g.Expect(*summary.Status.DiagnosticAccuracy).To(BeNumerically("~", 1.0, 0.001))
 		}, 30*time.Second).Should(Succeed())
